reorganize components and make a template, break out the changelog to its own file / page, and delete old templates
