+++ components/ScriptAdder/index.js
