+++ components/ScriptAdder/scripts.js
@@ -8,7 +8,7 @@ export const TOCBOT_INIT = {
   tocbot.init({
     tocSelector: '.js-toc',
     contentSelector: '.js-toc-content',
-    headingSelector: 'h1, h2, h3',
+    headingSelector: 'h1, h2, h3, h4',
     positionFixedSelector: '.js-toc',
     smoothScrollOptions: {
       easing: 'easeInOutCubic',
