+++ package.json
@@ -13,20 +13,20 @@
     "build-js": "webpack src/js/index.js static/js/tocbot.js -d",
     "build-js-watch": "webpack src/js/index.js static/js/tocbot.js -d --watch",
     "build-js-uglify": "webpack src/js/index.js static/js/tocbot.min.js -d --optimize-minimize",
+    "build-content": "processto \"*.md\" --outputDir pages --filenamePrefix _",
     "estimate-js-size": "gzip -9 -c static/js/tocbot.min.js | wc -c",
     "estimate-css-size": "gzip -9 -c static/css/tocbot.css | wc -c",
     "lint": "",
     "dev": "parallelshell \"npm run next\" \"npm run build-js-watch\" \"npm run build-css-watch\"",
     "start": "npm run build-css && npm run dev",
     "deploy": "npm run build && gh-pages -d build",
-    "test": "mocha --grep \"$TEST_NAME\" test/index.js",
-    "test:debug": "node-debug _mocha --timeout 0 --grep \"$TEST_NAME\" test/index.js",
+    "test": "mocha test/index.js",
+    "test:debug": "node-debug _mocha --timeout 0 test/index.js",
     "test:watch": "mocha --watch test/index.js",
     "next": "next",
     "next-build": "next build",
     "next-export": "next export",
-    "next-start": "next start",
-    "process-content": "processto \"*.md\" --outputDir pages --filenamePrefix _"
+    "next-start": "next start"
   },
   "repository": {
     "type": "git",
