+++ pages/changelog.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import Template from '../components/Template'
 
-import PAGE_JSON from './_README.json'
+import PAGE_JSON from './_CHANGELOG.json'
 const buttonClasses = 'f6 link dim br1 ph3 pv2 mb2 dib white bg-dark-gray'
 
 const Index = (props) => {
