--- pages/index.css
@@ -1,47 +1,0 @@
-/* Page Styles */
-
-.transition--300 {
-  transition: all 300ms ease-in-out;
-}
-
-.toc {
-  height: 100%;
-  width: 280px;
-  transform: translateX(0);
-}
-
-.content {
-  // margin-right: 280px;
-}
-
-.hero {
-  background-color: #54BC4B;
-}
-
-.title {
-  font-size: 3em;
-}
-
-.content {
-  margin-bottom: 95vh;
-}
-
-.content ul,
-.content ol {
-  list-style: inherit;
-}
-
-.content a {
-  color: #0977c3;
-  text-decoration: none;
-  border-bottom: 1px solid #EEE;
-  transition: all 300ms ease;
-}
-
-.content a:hover {
-  border-bottom: 1px solid #0977c3;
-}
-
-a.toc-link {
-  text-decoration: none;
-}
