+++ src/scss/_page-styles.scss
@@ -12,6 +12,10 @@
 
 .content {
   // margin-right: 280px;
+  h1, h2 {
+    padding-top: 0;
+    margin-top: 0;
+  }
 }
 
 .hero {
