--- src/templates/wrapper.html
@@ -1,70 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta http-equiv="x-ua-compatible" content="ie=edge">
-    <title>Tocbot</title>
-    <meta name="description" content="">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <!-- Place favicon.ico in the root directory -->
-    <link rel="stylesheet" href="https://unpkg.com/tachyons@4.7.0/css/tachyons.min.css"/>
-    <link rel="stylesheet" href="assets/css/styles.css">
-    <link rel="stylesheet" href="assets/css/tocbot.css">
-
-    <!-- <link rel="stylesheet" href="/test/mocha.css">
-    <script src="/test/mocha.js"></script> -->
-  </head>
-  <body>
-    <!-- APP_CONTENT_HERE -->
-
-    <script src="assets/js/tocbot.js"></script>
-    <script>
-      tocbot.init({
-        tocSelector: '.js-toc',
-        contentSelector: '.js-toc-content',
-        headingSelector: 'h1, h2, h3',
-        positionFixedSelector: '.js-toc',
-        smoothScrollOptions: {
-          easing: 'easeInOutCubic',
-          offset: 0,
-          callback: function(anchor, toggle) { console.log(anchor, toggle) },
-          speed: 300 // animation duration.
-        },
-      });
-    </script>
-
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/marked/0.3.5/marked.js"></script>
-    <script>
-      // Try it out
-      var markdownEl = document.querySelector('#try-it-markdown');
-      var tryItReset = document.querySelector('#try-it-reset');
-      var tryItCheckbox = document.querySelector('#try-it-checkbox');
-      var contentEl = document.querySelector('.js-toc-content');
-
-      var contentBackup = document.createElement('div');
-      contentBackup.id = 'content-backup';
-      contentBackup.innerHTML = contentEl.innerHTML;
-
-      markdownEl.addEventListener('change', function(event) {
-        var markdown = event.target.value;
-        contentEl.innerHTML = marked(markdown);
-        tocbot.refresh();
-      });
-
-      tryItReset.addEventListener('click', function(event) {
-        tryItCheckbox.checked = false;
-        contentEl.innerHTML = contentBackup.innerHTML;
-        tocbot.refresh();
-      });
-    </script>
-
-    <script>
-        (function(b,o,i,l,e,r){b.GoogleAnalyticsObject=l;b[l]||(b[l]=
-        function(){(b[l].q=b[l].q||[]).push(arguments)});b[l].l=+new Date;
-        e=o.createElement(i);r=o.getElementsByTagName(i)[0];
-        e.src='https://www.google-analytics.com/analytics.js';
-        r.parentNode.insertBefore(e,r)}(window,document,'script','ga'));
-        ga('create','UA-76620957-1','auto');ga('send','pageview');
-    </script>
-  </body>
-</html>
